requirements.txt
icrawler
pillow
imagehash
opencv-python
ultralytics
bing-image-downloader